WARNING: big rebase that rewrites remote git history from there
This shouldn't be done, but as I'm the only one using the repo I'm not screwing up others local git history with the repo. I'm doing it by laziness and to save some time, because I was doing a refactoring involving 3 branches and it was the easiest lazy way to do a big rebase.
I will try to make small commits in the rebase to follow the project's progression, but all the steps in the original history won't be there. But, well, it doesn't really matter.